Finalized undo and redo functionality
Edited the erase method to save the state of an item in the undo action. This includes the text, the icon, and the enabled status. Added a re-add method in the itemlist class that adds an item into the list at a given position in the list, with text, icon, and embedded status passed as arguments.

After adding an item in the editor and closing the editor, pressing ctrl+z or selecting removes the item. Pressing ctrl+shift+z adds the item back in. After adding more items, giving them new text and new icons, while arbitrarily changing the enabled status, and then deleting any element, close the editor window and pressing ctrl+z brings the item back into its position, with the changes intact. Pressing ctrl+shift+z removes the item again.